Ignore-this: a92cf47b8ebad04031159f5673bc5b65 NEW MAKEFILE! No more packs!
darcs-hash:20100629111803-e782f-61327dd4611277654d13858b5ce01badd79f1317

+++ trunk/goog/timer/timer.ml
@@ -14,10 +14,10 @@ open Js
 class type timer = object
   inherit eventTarget
   (** Gets the interval of the timer. *)
-  method getInterval : float meth
+  method getInterval : int meth
 
  (** Sets the interval of the timer.*)
-  method setInterval : float -> unit meth
+  method setInterval : int -> unit meth
 
  (** Dispatches the TICK event. This is its own method so subclasses can override. *)
   method dispatchTick : unit meth
@@ -33,5 +33,5 @@ class type timer = object
 
 end
 
-let timer : (float  -> timer t) constr =
+let timer : (int  -> timer t) constr =
   Js.Unsafe.variable "goog.Timer"
